+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GoogleSubscriber.scala
@@ -2,7 +2,7 @@ package org.broadinstitute.dsde.workbench.google2
 
 import cats.effect._
 import fs2.Stream
-import io.chrisdavenport.log4cats.StructuredLogger
+import org.typelevel.log4cats.StructuredLogger
 import io.circe.Decoder
 
 trait GoogleSubscriber[F[_], A] {