+++ util2/src/main/scala/org/broadinstitute/dsde/workbench/util2/syntax.scala
@@ -1,5 +1,5 @@
 package org.broadinstitute.dsde.workbench.util2
-import io.chrisdavenport.log4cats.Logger
+import org.typelevel.log4cats.Logger
 
 object syntax {
   implicit def logSyntax[F[_]](log: Logger[F]): ContextLogger[F] = ContextLogger(log)